requirements.txt
robotframework
robotframework-browser
robotframework-pabot 
robotframework-requests
PyYAML